separate settings and menu
