--- src/components/GameMenu/components/GameControl/GameControlStyle.ts
@@ -1,14 +1,0 @@
-import { css } from "@emotion/core";
-export const LayerStyle = css`
-  .Button {
-    margin: 0 5px;
-  }
-  .Button:first-of-type {
-    margin-left: 0px;
-  }
-  .Button:last-of-type {
-    margin-right: 0px;
-  }
-  border-radius: 0;
-  box-shadow: none;
-`;
