--- src/components/GameMenu/components/GameControl/index.ts
@@ -1,1 +1,0 @@
-export * from "./GameControl";
