--- src/components/GameMenu/components/PercentFilled/index.ts
@@ -1,1 +1,0 @@
-export * from "./PercentFilled";
