--- src/components/GameMenu/components/SizeForm/SizeForm.tsx
@@ -1,76 +1,0 @@
-import React, {
-  FunctionComponent,
-  ChangeEvent,
-  useState,
-  useEffect,
-} from "react";
-import { useTheme, InputGroup, Container } from "sancho";
-
-export interface SizeFormProps {
-  minWidth: number;
-  maxWidth: number;
-  minHeight: number;
-  maxHeight: number;
-  passSize: (values?: { width: number; height: number }) => void;
-}
-
-const SizeForm: FunctionComponent<SizeFormProps> = ({
-  minWidth = 3,
-  maxWidth = 10,
-  minHeight = 3,
-  maxHeight = 10,
-  passSize = (): void => {
-    console.log("pass size");
-  },
-}: SizeFormProps) => {
-  /* const { minWidth, maxWidth, minHeight, maxHeight, passSize } = props; */
-  const theme = useTheme();
-  const [size, setSize] = useState<{ width: number; height: number }>({
-    width: minWidth,
-    height: minHeight,
-  });
-  const changeSize = (e: ChangeEvent<HTMLInputElement>): void => {
-    const newState = { ...size, [e.target.name]: parseInt(e.target.value, 10) };
-    setSize(newState);
-  };
-  useEffect(() => {
-    passSize(size);
-  }, [size]);
-
-  return (
-    <Container css={{ background: theme.colors.background.layer, padding: 0 }}>
-      <form>
-        <InputGroup
-          label="Field width"
-          css={{ flexDirection: "column", display: "flex" }}
-        >
-          <input
-            value={size.width}
-            onChange={changeSize}
-            type="range"
-            min={minWidth}
-            max={maxWidth}
-            name="width"
-            step="1"
-          />
-        </InputGroup>
-        <InputGroup
-          label="Field height"
-          css={{ flexDirection: "column", display: "flex" }}
-        >
-          <input
-            value={size.height}
-            onChange={changeSize}
-            type="range"
-            min={minHeight}
-            max={maxHeight}
-            step="1"
-            name="height"
-          />
-        </InputGroup>
-      </form>
-    </Container>
-  );
-};
-
-export { SizeForm };
