--- src/components/GameMenu/components/SizeForm/index.ts
@@ -1,1 +1,0 @@
-export * from "./SizeForm";
