--- src/components/GameMenu/components/index.ts
@@ -1,3 +1,0 @@
-export * from "./PercentFilled";
-export * from "./SizeForm";
-export * from "./GameControl";
