--- src/components/GameMenu/index.ts
@@ -1,2 +1,0 @@
-export * from "./GameMenu";
-export * from "./components";
