+++ src/components/GameSettings/index.ts
