+++ src/rdx/reducder.ts
@@ -5,7 +5,7 @@ import {
 } from "@/features/lesson17/duck/reducer";
 import { userSlice } from "@/features/login/duck/reducer";
 import { appSlice } from "@/features/App/duck/reducer";
-import { settingsSlice } from "@/modules/GameSettings/duck/reducer";
+import { settingsSlice } from "@/modules/GameMenu/duck/reducer";
 import { gameSlice } from "@/modules/Game/duck/reducer";
 import { fieldSlice } from "@/modules/GameField/duck/reducer";
